+++ src/app/(auth)/verify-email/page.tsx
@@ -1,10 +1,21 @@
 import VerifyEmailForm from "@/components/verify-email-form";
-import React from "react";
+import React, { Suspense } from "react";
+
+// Add a loading component
+const Loading = () => {
+  return (
+    <div className="flex justify-center items-center">
+      <p>Loading verification...</p>
+    </div>
+  );
+};
 
 const VerifyEmailPage = () => {
   return (
     <div className="flex justify-center items-center h-screen">
-      <VerifyEmailForm />
+      <Suspense fallback={<Loading />}>
+        <VerifyEmailForm />
+      </Suspense>
     </div>
   );
 };
